app/requirements.txt
fastapi
uvicorn
cassandra-driver
pydantic[dotenv]
watchfiles